requirements.txt
python-telegram-bot
cowsay
nest_asyncio
python-dotenv